Consolidate payload cache function into workload manager
